requirements.txt
spotipy
pandas
python-dotenv
scikit-learn